BDF improvements 1. supports charactor offset, encorrect positioning: 2. supports more parameter, like DWIDTH 3. allow developer marks imported font( in case of user font not covered    whole BMP so unifont fallback trigged ) TODO 1. merge offset info to pixels, avoid additional memory usage( about 100KB ) 2. support FULL BDF specification

+++ text.h
@@ -201,9 +201,9 @@ INT
 PAL_MultiByteToWideCharCP(
 	CODEPAGE      cp,
 	LPCSTR        mbs,
-	int           mbslength,
+	size_t        mbslength,
 	LPWSTR        wcs,
-	int           wcslength
+	size_t        wcslength
 	);
 
 WCHAR
@@ -224,7 +224,7 @@ PAL_SetCodePage(
 CODEPAGE
 PAL_DetectCodePageForString(
 	const char *   text,
-	int            text_len,
+	size_t         text_len,
 	CODEPAGE       default_cp,
 	int *          probability
 );
